+++ libs/platform/src/lib/search-field/search-field.component.spec.ts
@@ -7,8 +7,7 @@ import '@angular/localize/init';
 import { Observable, of } from 'rxjs';
 
 import { RtlService } from '@fundamental-ngx/core/utils';
-import { createKeyboardEvent } from '@fundamental-ngx/platform/shared';
-import { DataProvider, SearchFieldDataSource } from '@fundamental-ngx/platform/shared';
+import { createKeyboardEvent, DataProvider, SearchFieldDataSource } from '@fundamental-ngx/platform/shared';
 import { PlatformSearchFieldModule } from './search-field.module';
 import { SearchFieldComponent, SearchInput, SuggestionItem, ValueLabelItem } from './search-field.component';
 
@@ -83,22 +82,20 @@ function mouseClickOnElement(el: Element): void {
 })
 class TestComponent {
     @ViewChild(SearchFieldComponent, { static: true }) component: SearchFieldComponent;
-    public placeholder: string;
-    public suggestions: SuggestionItem[] | Observable<SuggestionItem[]>;
-    public categories: ValueLabelItem[];
-    public categoryLabel: string;
-    public hideCategoryLabel = false;
-    public contentDensity: 'cozy' | 'compact';
-    public isLoading = false;
-    public disabled = false;
+    placeholder: string;
+    suggestions: SuggestionItem[] | Observable<SuggestionItem[]>;
+    categories: ValueLabelItem[];
+    categoryLabel: string;
+    hideCategoryLabel = false;
+    contentDensity: 'cozy' | 'compact';
+    isLoading = false;
+    disabled = false;
 
-    public inputValue: SearchInput;
-    public submitValue: SearchInput;
-    public isSearchCanceled = false;
+    inputValue: SearchInput;
+    submitValue: SearchInput;
+    isSearchCanceled = false;
 
     @ViewChild('outsideButton') outsideButton: ElementRef<HTMLElement>;
-
-    constructor() {}
 
     onInputChange($event): void {
         this.inputValue = $event;
@@ -240,8 +237,7 @@ describe('SearchFieldComponent', () => {
         expect(menuEls.length).toBe(0);
     });
 
-    // TODO: Upskip after fix
-    xit('should show the "category button" if "categories" is set with one or more items', () => {
+    it('should show the "category button" if "categories" is set with one or more items', () => {
         host.placeholder = 'Search';
         host.suggestions = [{ value: 'Apple' }, { value: 'Banana' }, { value: 'Carrot' }];
         host.categories = CATEGORIES;
@@ -252,11 +248,10 @@ describe('SearchFieldComponent', () => {
         expect(categoryButton.length).toBe(1);
 
         const categoryLabel = fixture.debugElement.query(By.css('.fdp-search-field__category-label'));
-        expect(categoryLabel.nativeElement.textContent).toBe('Category');
+        expect(categoryLabel.nativeElement.innerText).toBe('Category');
     });
 
-    // TODO: Unskip after fix
-    xit('should allow the user to set the text of the category label', () => {
+    it('should allow the user to set the text of the category label', () => {
         host.placeholder = 'Search';
         host.suggestions = [{ value: 'Apple' }, { value: 'Banana' }, { value: 'Carrot' }];
         host.categories = CATEGORIES;
@@ -264,7 +259,7 @@ describe('SearchFieldComponent', () => {
         fixture.detectChanges();
 
         const categoryLabel = fixture.debugElement.query(By.css('.fdp-search-field__category-label'));
-        expect(categoryLabel.nativeElement.textContent).toBe('Categoría');
+        expect(categoryLabel.nativeElement.innerText).toBe('Categoría');
     });
 
     it('should allow the user to hide the category label', () => {
@@ -289,8 +284,7 @@ describe('SearchFieldComponent', () => {
         expect(categoryDropdown.length).toBe(0);
     });
 
-    // TODO: Unskip after test
-    xit('should change the category label to the selected category', fakeAsync(() => {
+    it('should change the category label to the selected category', fakeAsync(() => {
         host.placeholder = 'Search';
         host.suggestions = [{ value: 'Apple' }, { value: 'Banana' }, { value: 'Carrot' }];
         host.categories = CATEGORIES;
@@ -309,9 +303,9 @@ describe('SearchFieldComponent', () => {
         (items[2] as HTMLElement).click();
         fixture.detectChanges();
 
-        expect(component.currentCategory).toEqual(CATEGORIES[2]);
+        expect(component._currentCategory).toEqual(CATEGORIES[2]);
         let categoryLabel = fixture.debugElement.query(By.css('.fdp-search-field__category-label'));
-        expect(categoryLabel.nativeElement.textContent).toBe(CATEGORIES[2].label);
+        expect(categoryLabel.nativeElement.innerText).toBe(CATEGORIES[2].label);
         expect(host.inputValue.category).toBe(CATEGORIES[2].value);
 
         // click on category button
@@ -325,9 +319,9 @@ describe('SearchFieldComponent', () => {
         (items[1] as HTMLElement).click();
         fixture.detectChanges();
 
-        expect(component.currentCategory).toEqual(CATEGORIES[1]);
+        expect(component._currentCategory).toEqual(CATEGORIES[1]);
         categoryLabel = fixture.debugElement.query(By.css('.fdp-search-field__category-label'));
-        expect(categoryLabel.nativeElement.textContent).toBe(CATEGORIES[1].label);
+        expect(categoryLabel.nativeElement.innerText).toBe(CATEGORIES[1].label);
         expect(host.inputValue.category).toBe(CATEGORIES[1].value);
     }));
 
@@ -379,7 +373,7 @@ describe('SearchFieldComponent', () => {
 
         menuEls = overlayContainerEl.querySelectorAll('.fd-menu');
         expect(menuEls.length).toBe(1);
-        expect(component.showDropdown).toBeTruthy();
+        expect(component._showDropdown).toBeTruthy();
     });
 
     it('should not create multiple overlays with subsequent keyboard entries', () => {
@@ -401,7 +395,7 @@ describe('SearchFieldComponent', () => {
 
         menuEls = overlayContainerEl.querySelectorAll('.fd-menu');
         expect(menuEls.length).toBe(1);
-        expect(component.showDropdown).toBeTruthy();
+        expect(component._showDropdown).toBeTruthy();
 
         textInput.nativeElement.value = 'ap';
         textInput.nativeElement.dispatchEvent(new Event('input'));
@@ -410,7 +404,7 @@ describe('SearchFieldComponent', () => {
 
         menuEls = overlayContainerEl.querySelectorAll('.fd-menu');
         expect(menuEls.length).toBe(1);
-        expect(component.showDropdown).toBeTruthy();
+        expect(component._showDropdown).toBeTruthy();
     });
 
     it('should set input text and close dropdown on select of item', () => {
@@ -442,7 +436,7 @@ describe('SearchFieldComponent', () => {
         // check to see if dropdown is open
         menuEls = overlayContainerEl.querySelectorAll('.fd-menu');
         expect(menuEls.length).toBe(0);
-        expect(component.showDropdown).toBeFalsy();
+        expect(component._showDropdown).toBeFalsy();
 
         // check input text
         expect(component.inputText).toBe('Apple');
@@ -561,7 +555,7 @@ describe('SearchFieldComponent', () => {
 
         const menuEls = overlayContainerEl.querySelectorAll('.fd-menu');
         expect(menuEls.length).toBe(0);
-        expect(component.showDropdown).toBeFalsy();
+        expect(component._showDropdown).toBeFalsy();
     });
 
     it('should be able to be put into "isLoading" state', () => {
@@ -649,7 +643,7 @@ describe('SearchFieldComponent', () => {
         // check dropdown
         const menuEls = overlayContainerEl.querySelectorAll('.fd-menu');
         expect(menuEls.length).toBe(0);
-        expect(component.showDropdown).toBeFalsy();
+        expect(component._showDropdown).toBeFalsy();
 
         expect(host.inputValue).toEqual({ text: '', category: null });
     });
@@ -681,11 +675,11 @@ describe('SearchFieldComponent', () => {
         // check dropdown
         const menuEls = overlayContainerEl.querySelectorAll('.fd-menu');
         expect(menuEls.length).toBe(0);
-        expect(component.showDropdown).toBeFalsy();
+        expect(component._showDropdown).toBeFalsy();
     });
 
     // TODO: Unskip after fix
-    xit('should return focus to the input field after suggestion item is selected', () => {
+    it('should return focus to the input field after suggestion item is selected', () => {
         // set up component
         host.placeholder = 'Search';
         host.suggestions = [{ value: 'Apple' }, { value: 'Banana' }, { value: 'Carrot' }];
@@ -728,7 +722,7 @@ describe('SearchFieldComponent', () => {
         // check dropdown
         let menuEls = overlayContainerEl.querySelectorAll('.fd-menu');
         expect(menuEls.length).toBe(1);
-        expect(component.showDropdown).toBeTruthy();
+        expect(component._showDropdown).toBeTruthy();
 
         /// click outside
         host.outsideButton.nativeElement.click();
@@ -736,7 +730,7 @@ describe('SearchFieldComponent', () => {
         // check dropdown
         menuEls = overlayContainerEl.querySelectorAll('.fd-menu');
         expect(menuEls.length).toBe(0);
-        expect(component.showDropdown).toBeFalsy();
+        expect(component._showDropdown).toBeFalsy();
     });
 });
 
@@ -763,18 +757,18 @@ describe('SearchFieldComponent', () => {
 })
 class DataSourceTestComponent implements OnInit {
     @ViewChild(SearchFieldComponent, { static: true }) component: SearchFieldComponent;
-    public placeholder: string;
-    public categories: ValueLabelItem[];
-    public categoryLabel: string;
-    public hideCategoryLabel = false;
-    public contentDensity: 'cozy' | 'compact';
-    public isLoading = false;
-    public disabled = false;
-    public dataSource: SearchFieldDataSource<any>;
+    placeholder: string;
+    categories: ValueLabelItem[];
+    categoryLabel: string;
+    hideCategoryLabel = false;
+    contentDensity: 'cozy' | 'compact';
+    isLoading = false;
+    disabled = false;
+    dataSource: SearchFieldDataSource<any>;
 
-    public inputValue: SearchInput;
-    public submitValue: SearchInput;
-    public isSearchCanceled = false;
+    inputValue: SearchInput;
+    submitValue: SearchInput;
+    isSearchCanceled = false;
 
     @ViewChild('outsideButton') outsideButton: ElementRef<HTMLElement>;
 
@@ -839,7 +833,7 @@ describe('SearchFieldComponent with DataSource', () => {
         // check dropdown
         let menuEls = overlayContainerEl.querySelectorAll('.fd-menu');
         expect(menuEls.length).toBe(1);
-        expect(component.showDropdown).toBeTruthy();
+        expect(component._showDropdown).toBeTruthy();
         let items = getDropdownItems(menuEls[0]);
         expect(items.length).toBe(1);
         expect(items[0].textContent).toBe('Apple');
@@ -852,7 +846,7 @@ describe('SearchFieldComponent with DataSource', () => {
         // check dropdown
         menuEls = overlayContainerEl.querySelectorAll('.fd-menu');
         expect(menuEls.length).toBe(1);
-        expect(component.showDropdown).toBeTruthy();
+        expect(component._showDropdown).toBeTruthy();
         items = getDropdownItems(menuEls[0]);
         expect(items.length).toBe(2);
         expect(items[0].textContent).toBe('Banana');
@@ -874,7 +868,7 @@ describe('SearchFieldComponent with DataSource', () => {
         // check dropdown
         const menuEls = overlayContainerEl.querySelectorAll('.fd-menu');
         expect(menuEls.length).toBe(1);
-        expect(component.showDropdown).toBeTruthy();
+        expect(component._showDropdown).toBeTruthy();
         const items = getDropdownItems(menuEls[0]);
         expect(items.length).toBe(2);
         expect(items[0].textContent).toBe('Almond');